Refactoring to one class per file
