+++ SpectralAveraging/NoiseEstimates/WaveletEnums.cs
@@ -1,0 +1,11 @@
+﻿namespace SpectralAveraging.NoiseEstimates;
+
+public enum WaveletType
+{
+    Haar = 1
+}
+
+public enum BoundaryType
+{
+    Reflection = 1
+}
