+++ SpectralAveraging/NoiseEstimates/WaveletMath.cs
@@ -111,150 +111,4 @@ namespace SpectralAveraging.NoiseEstimates
             return ModWt(signal, filters.WaveletCoefficients, filters.ScalingCoefficients, filters.WaveletType);
         }
     }
-
-    public enum WaveletType
-    {
-        Haar = 1
-    }
-
-    public enum BoundaryType
-    {
-        Reflection = 1
-    }
-
-    public class ModWtOutput
-    {
-        public ModWtOutput(int maxScale, WaveletType waveletType, BoundaryType boundaryType)
-        {
-            Levels = new List<Level>(); 
-            MaxScale = maxScale;
-            WaveletType = waveletType;
-            BoundaryType = boundaryType; 
-        }
-
-        public List<Level> Levels { get; private set; }
-        public int MaxScale { get; private set; }
-        public WaveletType WaveletType { get; }
-        public BoundaryType BoundaryType { get; }
-        public void AddLevel(Level level)
-        {
-            Levels.Add(level);
-        }
-
-        public void AddLevel(double[] waveletCoeff, double[] scalingCoeff, int scale, 
-            BoundaryType boundaryType, int originalSignalLength, int filterLength)
-        {
-            if (boundaryType == BoundaryType.Reflection)
-            {
-                int startIndex = ((int)Math.Pow(2, scale))*(filterLength-1);
-                int stopIndex = startIndex + originalSignalLength; 
-                Levels.Add(new Level(scale, 
-                    waveletCoeff[startIndex .. stopIndex], 
-                    scalingCoeff[startIndex .. stopIndex]));
-            }
-        }
-
-        public void PrintToTxt(string path)
-        {
-            StringBuilder sb = new();
-            foreach (var level in Levels)
-            {
-                sb.AppendLine(string.Join("\t", level.WaveletCoeff)); 
-                sb.AppendLine(string.Join("\t", level.ScalingCoeff));
-            }
-            File.WriteAllText(path, sb.ToString());
-        }
-    }
-
-    public class Level
-    {
-        public Level(int scale, double[] waveletCoeff, double[] scalingCoeff)
-        {
-            Scale = scale; 
-            WaveletCoeff = waveletCoeff;
-            ScalingCoeff = scalingCoeff;
-        }
-
-        public Level(int scale)
-        {
-            Scale = scale; 
-        }
-
-        public int Scale { get; private set; } 
-        public double[] WaveletCoeff { get; private set; }
-        public double[] ScalingCoeff { get; private set; }
-
-    }
-
-    public class WaveletFilter
-    {
-        public double[] WaveletCoefficients { get; private set; }
-        public double[] ScalingCoefficients { get; private set; }
-        public WaveletType WaveletType { get; private set; }
-
-        public void CreateFiltersFromCoeffs(double[] filterCoeffs)
-        {
-            WaveletCoefficients = new double[filterCoeffs.Length];
-            ScalingCoefficients = new double[filterCoeffs.Length]; 
-
-            // calculate wavelet coefficients
-            for (int i = 0; i < ScalingCoefficients.Length; i++)
-            {
-                ScalingCoefficients[i] = filterCoeffs[i] / Math.Sqrt(2d); 
-            }
-            WaveletCoefficients = WaveletMathUtils.QMF(ScalingCoefficients, inverse: true); 
-        }
-
-        public void CreateFiltersFromCoeffs(WaveletType waveletType)
-        {
-            switch (waveletType)
-            {
-                case WaveletType.Haar:
-                {
-                    WaveletType = WaveletType.Haar; 
-                    CreateFiltersFromCoeffs(_haarCoefficients);
-                    return; 
-                }
-            }
-        }
-        private readonly double[] _haarCoefficients =
-        {
-            0.7071067811865475,
-            0.7071067811865475
-        }; 
-    }
-
-    public static class WaveletMathUtils
-    {
-        /// <summary>
-        /// Calcualte the quadruture mirror filter for x, an array of filter coefficients
-        /// </summary>
-        /// <param name="x"></param>
-        /// <param name="inverse"></param>
-        /// <returns></returns>
-        public static double[] QMF(double[] x, bool inverse = false)
-        {
-            double[] y = new double[x.Length];
-            Buffer.BlockCopy(x, 0, y, 0, x.Length * sizeof(double));
-            Array.Reverse(y);
-            if (inverse)
-            {
-                // start the for loop from the back
-                int firstIndex = 1; 
-                for (int i = x.Length - 1; i >= 0; i--)
-                {
-                    y[i] *= Math.Pow(-1d, firstIndex);
-                    firstIndex++;
-                }
-            }
-            else
-            {
-                for (int i = 0; i < x.Length; i++)
-                {
-                    y[i] *= Math.Pow(-1d, i + 1);
-                }
-            }
-            return y; 
-        }
-    }
 }
